Track balances as Number64UintType (#2973)
* Track balances as Number64UintType

* Use Number64ListType ssz type, refactor BalanceList

* Use applyDeltaInBatch

* Fix spec & unit tests

* Fix perf test to go with numbered balance

* New ssz, persistent-merkle-tree and fix tests

* Chore: clean up commented code

* Fix test/perf/phase0/epoch/processEffectiveBalanceUpdates.test.ts

+++ packages/light-client/test/lightclientMockServer.ts
@@ -95,7 +95,7 @@ export class LightclientMockServer {
     // Increase balances, simulate rewards
     if (slot % SLOTS_PER_EPOCH === 0) {
       for (let i = 0, len = state.balances.length; i < len; i++) {
-        state.balances[i] = state.balances[i] + BigInt(Math.round(11430 * (1 + Math.random())));
+        state.balances[i] = state.balances[i] + Math.round(11430 * (1 + Math.random()));
       }
     }
